This MR makes 3 changes to improve the layer layouts:
- calls to `g.Layers()` are now stored in a variable to avoid computing it in loops (layer assigner, cycle remover, brandeskopf)
- neighbors are also cached to avoid parsing the whole segment list when trying to find neighbors of a single node (brandeskopf, cycle_remover)
- crossings comparisons for adjacent nodes in a layer are now computed only between these 2 adjacent nodes, since it is not useful to add crossing of unrelated nodes


Benchmarks


Running with No ordering assigner (0 epoch)
```go
	gl := layout.SugiyamaLayersStrategyGraphLayout{
		CycleRemover:   layout.NewSimpleCycleRemover(),
		LevelsAssigner: layout.NewLayeredGraph,
		OrderingAssigner: layout.WarfieldOrderingOptimizer{
			Epochs:                   0,
			LayerOrderingInitializer: layout.BFSOrderingInitializer{},
			LayerOrderingOptimizer: layout.CompositeLayerOrderingOptimizer{
				Optimizers: []layout.LayerOrderingOptimizer{
					// layout.WMedianOrderingOptimizer{},
					layout.SwitchAdjacentOrderingOptimizer{},
				},
			},
		}.Optimize,
		NodesHorizontalCoordinatesAssigner: layout.BrandesKopfLayersNodesHorizontalAssigner{
			Delta: 125,
		},
		NodesVerticalCoordinatesAssigner: layout.BasicNodesVerticalCoordinatesAssigner{
			MarginLayers:   125,
			FakeNodeHeight: 50,
		},
		EdgePathAssigner: layout.StraightEdgePathAssigner{}.UpdateGraphLayout,
	}
	gl.UpdateGraphLayout(g)

```
Example: tree with 5000 nodes:
[input.jsonl.json](https://github.com/user-attachments/files/17965677/input.jsonl.json)

Time for main version: 23 seconds
cycle remover:  322.1371ms
levels assigner:  328.6456ms
validate:  0s
ordering assigner:  2.5544ms
node horizontal coordinates:  22.2575877s
node vertical coordinates:  5.3355ms

Time for this version: 50ms seconds
cycle remover:  2.1821ms
levels assigner:  5.5674ms
validate:  536µs
ordering assigner:  4.5944ms
node horizontal coordinates:  22.4342ms
node vertical coordinates:  4.5977ms

Example: tree with 50000 nodes:
[input-50000.jsonl.json](https://github.com/user-attachments/files/17965746/input-50000.jsonl.json)

Time for main version: ?
cycle remover:  32.1171135s
levels assigner:  32.568738s
validate:  5.0201ms
ordering assigner:  46.0467ms
node horizonal coordinates: ? Didn't let it finish

Time for this version: 550ms
cycle remover:  37.5156ms
levels assigner:  48.2833ms
validate:  9.937ms
ordering assigner:  39.7217ms
node horizontal coordinates:  257.2831ms
node vertical coordinates:  52.993ms