@gverger 2024-12-02
Thanks for your comments, they make sense!

I can take over if you don't plan to go back at it. I like the direction it has taken. I don't know how long I will make changes to it, but I can take care of it a little bit :)