@nikolaydubina 2024-12-03
ok, I see you renamed. but I am still getting error. I guess no other way but full remove it. looks like GitHub trackes repos not by name but by some hashes/ids.

so that we don't miss your updates and this history, let's merge this into master first. then please back up locally your repo and all your branches. ping me when you delete it. and I will transfer it.

I am merging these PRs first!

<img width="699" alt="Screenshot 2024-12-03 at 09 20 15" src="https://github.com/user-attachments/assets/0b04c50f-13d8-4e0c-b30e-b371a14b6483">