@nikolaydubina 2024-12-03
yes, there should not be fork, even renamed one. https://docs.github.com/en/repositories/creating-and-managing-repositories/transferring-a-repository

<img width="801" alt="image" src="https://github.com/user-attachments/assets/d78bffe4-9a9e-4fef-9bb6-c398bf5eb0ee">